+++ inferno-types/src/Inferno/Types/Syntax.hs
@@ -40,6 +40,7 @@ module Inferno.Types.Syntax
     Pat (..),
     PatF (..),
     TV (..),
+    CustomType,
     BaseType (..),
     InfernoType (..),
     Expr
@@ -148,6 +149,9 @@ newtype TV = TV {unTV :: Int}
   deriving stock (Eq, Ord, Show, Data, Generic)
   deriving newtype (ToJSON, FromJSON, ToJSONKey, FromJSONKey, NFData, Hashable, Serialize)
 
+-- | Custom type names
+type CustomType = String
+
 data BaseType
   = TInt
   | TDouble
@@ -159,8 +163,10 @@ data BaseType
   | TTimeDiff
   | TResolution
   | TEnum Text (Set.Set Ident)
+  | TCustom CustomType
   deriving (Show, Eq, Ord, Data, Generic, ToJSON, FromJSON, NFData)
 
+-- Serialize is needed by VPDB to store runtime type reps
 instance Serialize BaseType where
   get =
     Serialize.getInt8 >>= \case
@@ -173,10 +179,14 @@ instance Serialize BaseType where
       6 -> pure TTime
       7 -> pure TTimeDiff
       8 -> pure TResolution
-      _ -> do
+      9 -> do
         nm <- Serialize.get
         ids <- Serialize.get
         pure $ TEnum (Text.decodeUtf8 nm) $ Set.fromList $ map (Ident . Text.decodeUtf8) ids
+      10 -> do
+        t <- Serialize.get
+        pure $ TCustom t
+      _ -> error "Unknown serialization of BaseType"
 
   put = \case
     TInt -> Serialize.putInt8 0
@@ -192,7 +202,11 @@ instance Serialize BaseType where
       Serialize.putInt8 9
       Serialize.put $ Text.encodeUtf8 nm
       Serialize.put $ map (Text.encodeUtf8 . unIdent) $ Set.toList ids
+    TCustom t -> do
+      Serialize.putInt8 10
+      Serialize.put t
 
+-- Hashable is needed by Haxl for requests
 instance Hashable BaseType where
   hashWithSalt s TInt = hashWithSalt s (1 :: Int)
   hashWithSalt s TDouble = hashWithSalt s (2 :: Int)
@@ -204,6 +218,7 @@ instance Hashable BaseType where
   hashWithSalt s TTimeDiff = hashWithSalt s (8 :: Int)
   hashWithSalt s TResolution = hashWithSalt s (9 :: Int)
   hashWithSalt s (TEnum nm cs) = hashWithSalt s (10 :: Int, nm, Set.toList cs)
+  hashWithSalt s (TCustom t) = hashWithSalt s (11 :: Int, t)
 
 data InfernoType
   = TVar TV
@@ -244,6 +259,7 @@ instance Pretty BaseType where
     TTimeDiff -> "timeDiff"
     TResolution -> "resolution"
     TEnum t _ -> pretty t
+    TCustom t -> pretty t
 
 instance Pretty InfernoType where
   pretty = \case